Update to rename dashboard files
