+++ frontend/app/dashboard/fit_score_chart.tsx
