+++ frontend/app/dashboard/improvement_suggestions.tsx
